stock reserving with tolerance and deducting from stock functions now working as expected

+++ app/Stocks/ProductionIngredientMove.php
@@ -9,7 +9,7 @@ class ProductionIngredientMove extends Moves
 {
     private $workOrder;
 
-    public function __construct(WorkOrder $workOrder, int $ingredientId, String $lotNumber, int $amount)
+    public function __construct(WorkOrder $workOrder, int $ingredientId, String $lotNumber, float $amount)
     {
         $this->workOrder = $workOrder;
         $this->type = 'production_ingredient';
